Refactor powder tower logic into dedicated module
